+++ routers/index.js
@@ -18,13 +18,19 @@ const [
   workflow,
   goodses,
   portal,
+  rdPortal,
+  adPortal,
+  psPortal,
   node
 ] = [
   require(resolve(__dirname, "user")),
   require(resolve(__dirname, "workflow")),
   require(resolve(__dirname, "goodses")),
   require(resolve(__dirname, "portal")),
-  require(resolve(__dirname, "nodeManager"))
+  require(resolve(__dirname, "rd-portal")),
+  require(resolve(__dirname, "ad-portal")),
+  require(resolve(__dirname, "ps-portal")),
+  require(resolve(__dirname, "nodeManager")),
 ];
 
 const [authentication] = [require(resolve(interceptors, "authentication"))];
@@ -46,15 +52,18 @@ module.exports = _ => {
   _.get("/portal/login", portal.login); // *页面跳转 -> 登录页
   _.get("/portal/home", authentication, portal.home); // *页面跳转 -> 首页
   _.get("/portal/workflow/:id", authentication, portal.info); // *页面跳转 -> 工作流详情页
-  // _.get("/portal/directory", authentication, portal.directory); // *页面跳转 -> 公共目录
 
   /** 研发部 **/
-  _.get("/portal/rd/home", authentication, portal.rdHome); // *页面跳转 -> 研发部首页
-  _.get("/portal/rd/platform", authentication, portal.platform); // *页面跳转 -> 平台
-  _.get("/portal/rd/product", authentication, portal.product); // *页面跳转 -> 产品
+  _.get("/portal/rd/home", authentication, rdPortal.home); // *页面跳转 -> 研发部首页
+  _.get("/portal/rd/platform", authentication, rdPortal.platform); // *页面跳转 -> 研发部 平台
+  _.get("/portal/rd/product", authentication, rdPortal.product); // *页面跳转 -> 产品
+  _.get("/portal/rd/directory", authentication, rdPortal.directory); // *页面跳转 -> 产品
 
-  _.get("/portal/ad/home", authentication, portal.adHome); // *页面跳转 -> 自主交付部首页
-  _.get("/portal/ps/home", authentication, portal.psHome); // *页面跳转 -> 售前技术部部首页
+  /** 自主交付部 **/
+  _.get("/portal/ad/home", authentication, adPortal.home); // *页面跳转 -> 自主交付部首页
+
+  /** 售前技术部 **/
+  _.get("/portal/ps/home", authentication, psPortal.home); // *页面跳转 -> 售前技术部部首页
 
 
   /*
@@ -68,14 +77,14 @@ module.exports = _ => {
    *  工作流模块
    */
   _.get("/workflow/list", authentication, workflow.list); // 工作流信息列表
-  _.get("//workflow/:id", authentication, workflow.detail);// 工作流详细信息
+  _.get("/workflow/:id", authentication, workflow.detail);// 工作流详细信息
   _.get("/workflow/:id/simple", authentication, workflow.simpleInfo); // 获取工作流实例简单信息
   _.get("/workflow/:nodeId/files", authentication, workflow.nodeFiles); // 获取工作流某实例下的文件列表
   _.get("/workflow/nodeInstance/:id", authentication, workflow.nodeInstance); // 获取工作流下实例节点信息
 
-  _.get("/workflow/overviews/rd/platform", authentication, workflow.rdOverviewsPlatform); // 总览 - 平台工作流
-  _.get("/workflow/overviews/rd/product", authentication, workflow.rdOverviewsProduct);// 总览 - 产品工作流
-  _.get("/workflow/overviews/rd/directory", authentication, workflow.rdOverviewsDirectory); // 总览 - 公共目录
+  _.get("/workflow/overviews/rd/platform", authentication, workflow.rdOverviewsPlatform); // 总览 - rd部门 平台预览
+  _.get("/workflow/overviews/rd/product", authentication, workflow.rdOverviewsProduct);// 总览 - rd部门 产品预览
+  _.get("/workflow/overviews/rd/directory", authentication, workflow.rdOverviewsDirectory); // 总览 - rd部门 公共目录预览
 
   /*
    *  文件系统模块